+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/adapter/AnnotationHandlerAdapter.java
@@ -1,10 +1,9 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe.adapter;
+package webmvc.org.springframework.web.servlet.mvc.adapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import webmvc.org.springframework.web.servlet.ModelAndView;
-import webmvc.org.springframework.web.servlet.mvc.tobe.HandlerExecution;
-import webmvc.org.springframework.web.servlet.mvc.tobe.adapter.HandlerAdapter;
+import webmvc.org.springframework.web.servlet.mvc.HandlerExecution;
 
 public class AnnotationHandlerAdapter implements HandlerAdapter {
     @Override
