+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/adapter/HandlerAdapter.java
@@ -1,4 +1,4 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe.adapter;
+package webmvc.org.springframework.web.servlet.mvc.adapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
