+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/mapper/HandlerMappingRegistry.java
@@ -1,10 +1,9 @@
-package webmvc.org.springframework.web.servlet.mvc.tobe.mapper;
+package webmvc.org.springframework.web.servlet.mvc.mapper;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import webmvc.org.springframework.web.servlet.mvc.tobe.mapper.HandlerMapping;
 
 public class HandlerMappingRegistry {
